requirements.txt
PyQt6
fpdf2
